@@ -1,7 +1,10 @@
 /** @type {import('tailwindcss').Config} */
-module.exports = {
+export default {
   content: ['./index.html', './src/**/*.{vue,js,ts,jsx,tsx}'],
   theme: {
+    fontFamily:{
+      greycliff :['"Greycliff CF"', 'sans-serif']
+    },
     extend: {}
   },
   plugins: []
