@@ -1,6 +1,16 @@
+import './assets/css/style.css'
+
 import { createApp } from 'vue'
 import App from './App.vue'
 
-import './index.css'
+import { createRouter, createWebHistory } from 'vue-router/auto'
 
-createApp(App).mount('#app')
+const app = createApp(App)
+
+app.use(
+  createRouter({
+    history: createWebHistory()
+  })
+)
+
+app.mount('#app')
