@@ -2,13 +2,21 @@ import { fileURLToPath, URL } from 'node:url'
 
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
+import VueDevTools from 'vite-plugin-vue-devtools'
+
+import VueRouter from 'unplugin-vue-router/vite'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [vue()],
+  plugins: [VueRouter({}), vue() /* ,VueDevTools() */],
   resolve: {
     alias: {
       '@': fileURLToPath(new URL('./src', import.meta.url))
     }
   }
+  // // décommenter pour build en mode production directement dans pocketbase
+  // ,
+  // build: {
+  //   outDir: './pocketbase/pb_public'
+  // }
 })
